requirements.txt
absl-py==1.4.0 ; python_version >= "3.10" and python_version < "4.0"
anyio==3.6.2 ; python_version >= "3.10" and python_version < "4.0"
cached-property==1.5.2 ; python_version >= "3.10" and python_version < "4.0"
certifi==2022.12.7 ; python_version >= "3.10" and python_version < "4"
charset-normalizer==3.0.1 ; python_version >= "3.10" and python_version < "4"
chex==0.1.6 ; python_version >= "3.10" and python_version < "4.0"
click==8.1.3 ; python_version >= "3.10" and python_version < "4.0"
colorama==0.4.6 ; python_version >= "3.10" and python_version < "4.0" and platform_system == "Windows"
contourpy==1.0.7 ; python_version >= "3.10" and python_version < "4.0"
cycler==0.11.0 ; python_version >= "3.10" and python_version < "4.0"
dm-tree==0.1.8 ; python_version >= "3.10" and python_version < "4.0"
etils==1.0.0 ; python_version >= "3.10" and python_version < "4.0"
fastapi==0.92.0 ; python_version >= "3.10" and python_version < "4.0"
filelock==3.9.0 ; python_version >= "3.10" and python_version < "4.0"
flax==0.6.6 ; python_version >= "3.10" and python_version < "4.0"
fonttools==4.39.0 ; python_version >= "3.10" and python_version < "4.0"
h11==0.14.0 ; python_version >= "3.10" and python_version < "4.0"
huggingface-hub==0.12.1 ; python_version >= "3.10" and python_version < "4.0"
idna==3.4 ; python_version >= "3.10" and python_version < "4"
importlib-resources==5.12.0 ; python_version >= "3.10" and python_version < "4.0"
jax==0.4.5 ; python_version >= "3.10" and python_version < "4.0"
jaxlib==0.4.4 ; python_version >= "3.10" and python_version < "4.0"
kiwisolver==1.4.4 ; python_version >= "3.10" and python_version < "4.0"
markdown-it-py==2.2.0 ; python_version >= "3.10" and python_version < "4.0"
matplotlib==3.7.1 ; python_version >= "3.10" and python_version < "4.0"
mdurl==0.1.2 ; python_version >= "3.10" and python_version < "4.0"
msgpack==1.0.4 ; python_version >= "3.10" and python_version < "4.0"
numpy==1.24.2 ; python_version >= "3.10" and python_version < "4.0"
opt-einsum==3.3.0 ; python_version >= "3.10" and python_version < "4.0"
optax==0.1.4 ; python_version >= "3.10" and python_version < "4.0"
orbax==0.1.3 ; python_version >= "3.10" and python_version < "4.0"
packaging==23.0 ; python_version >= "3.10" and python_version < "4.0"
pillow==9.4.0 ; python_version >= "3.10" and python_version < "4.0"
pydantic==1.10.5 ; python_version >= "3.10" and python_version < "4.0"
pygments==2.14.0 ; python_version >= "3.10" and python_version < "4.0"
pyparsing==3.0.9 ; python_version >= "3.10" and python_version < "4.0"
python-dateutil==2.8.2 ; python_version >= "3.10" and python_version < "4.0"
pyyaml==6.0 ; python_version >= "3.10" and python_version < "4.0"
regex==2022.10.31 ; python_version >= "3.10" and python_version < "4.0"
requests==2.28.2 ; python_version >= "3.10" and python_version < "4"
rich==13.3.2 ; python_version >= "3.10" and python_version < "4.0"
scipy==1.9.3 ; python_version >= "3.10" and python_version < "4.0"
six==1.16.0 ; python_version >= "3.10" and python_version < "4.0"
sniffio==1.3.0 ; python_version >= "3.10" and python_version < "4.0"
starlette==0.25.0 ; python_version >= "3.10" and python_version < "4.0"
tensorstore==0.1.33 ; python_version >= "3.10" and python_version < "4.0"
tokenizers==0.13.2 ; python_version >= "3.10" and python_version < "4.0"
toolz==0.12.0 ; python_version >= "3.10" and python_version < "4.0"
tqdm==4.64.1 ; python_version >= "3.10" and python_version < "4.0"
transformers==4.26.1 ; python_version >= "3.10" and python_version < "4.0"
typing-extensions==4.5.0 ; python_version >= "3.10" and python_version < "4.0"
urllib3==1.26.14 ; python_version >= "3.10" and python_version < "4"
uvicorn==0.20.0 ; python_version >= "3.10" and python_version < "4.0"
aws-cdk-lib==2.70.0
aws-cdk.asset-awscli-v1==2.2.121
aws-cdk.asset-kubectl-v20==2.1.1
aws-cdk.asset-node-proxy-agent-v5==2.0.98